+++ src/components/controls.ts
@@ -6,7 +6,6 @@ import { html } from "lit";
 import { Cell, ControlsDefinition } from "../types";
 import { getAvailablePropertyTypes, registry } from "../cellProperties/registry";
 import { renderIcon } from "./helpers/icon";
-
 // Note: These controls are not "Components" in the lit sense
 
 export function cellControlsTemplate(controls: ControlsDefinition) {
@@ -16,8 +15,10 @@ export function cellControlsTemplate(controls: ControlsDefinition) {
       (button) =>
         html`
           <button
-            @click="${button.callback}"
-            class="btn cell-controls-button py-1 ${button.hide === undefined ? "auto-hide" : button.hide} "
+            @click="${function (evt: Event) {
+              button.callback(evt);
+            }}"
+            class="btn cell-controls-button py-1"
             title="${button.tooltip}"
           >
             ${renderIcon(button.icon, { width: 16, height: 16 })}