+++ src/components/editor/codemirror/editor.ts
@@ -54,7 +54,7 @@ export function createCodeMirrorEditor(
     language?: string;
     wordWrap?: "off" | "on" | "wordWrapColumn" | "bounded";
   },
-  _runtime: Runtime
+  runtime: Runtime
 ) {
   const listen = EditorView.updateListener.of((update) => {
     if (update.docChanged) {
@@ -73,12 +73,7 @@ export function createCodeMirrorEditor(
           const firstLine = target.state.doc.line(1);
           const cursorPosition = target.state.selection.ranges[0].head;
           if (firstLine.from <= cursorPosition && cursorPosition <= firstLine.to) {
-            _runtime.controls.emit({
-              id: cell.id,
-              type: "FOCUS_CELL",
-              focus: "previous",
-            });
-            return true;
+            return runtime.controls.focusCell({ id: cell.id, focusTarget: "previous" });
           }
         }
         return false;
@@ -91,12 +86,7 @@ export function createCodeMirrorEditor(
           const lastline = target.state.doc.line(target.state.doc.lines);
           const cursorPosition = target.state.selection.ranges[0].head;
           if (lastline.from <= cursorPosition && cursorPosition <= lastline.to) {
-            _runtime.controls.emit({
-              id: cell.id,
-              type: "FOCUS_CELL",
-              focus: "next",
-            });
-            return true;
+            return runtime.controls.focusCell({ id: cell.id, focusTarget: "next" });
           }
         }
         return false;
@@ -126,7 +116,7 @@ export function createCodeMirrorEditor(
   };
 
   let isLocked: boolean | undefined = cell.metadata.properties.locked;
-  _runtime.controls.subscribeToCellChanges(cell.id, () => {
+  runtime.controls.subscribeToCellChanges(cell.id, () => {
     // Note this function will be called on ALL text changes, so any letter typed,
     // it's probably better for performance to only ask cm to change it's editable state if it actually changed.
     if (isLocked === cell.metadata.properties.locked) return;