+++ src/components/insertionLine.ts
@@ -8,7 +8,7 @@ import { customElement, query } from "lit/decorators.js";
 import { createPopper } from "@popperjs/core";
 import { CellTypePicker } from "./cellTypePicker";
 import { CellElement } from "./cell";
-import { Cell } from "../types";
+import { Cell, Runtime } from "../types";
 
 // Lazily initialized.. but cached for re-use.
 let globalCellTypePicker: CellTypePicker;
@@ -22,6 +22,7 @@ export class InsertionLine extends LitElement {
   hoverArea?: HTMLDivElement;
 
   private insertPosition: "before" | "after" = "after";
+  private runtime: Runtime;
 
   createRenderRoot() {
     return this;
@@ -29,6 +30,8 @@ export class InsertionLine extends LitElement {
 
   constructor() {
     super();
+    // TODO: pass this in..
+    this.runtime = window.runtime;
   }
 
   connectedCallback() {
@@ -81,10 +84,9 @@ export class InsertionLine extends LitElement {
         globalCellTypePicker.onInsert = (cellData: Partial<Cell>) => {
           // Right now we assume the insertion line has a cell as parent
           if (parent && parent instanceof CellElement) {
-            parent.runtime.controls.emit({
-              type: "INSERT_CELL",
+            this.runtime.controls.insertCell({
+              adjacentCellId: parent.cell.id,
               position: this.insertPosition,
-              id: parent.cell.id,
               data: cellData,
             });
             unpop();
@@ -98,10 +100,9 @@ export class InsertionLine extends LitElement {
   quickInsert(cellType: string) {
     const parent = this.parentElement;
     if (parent && parent instanceof CellElement) {
-      parent.runtime.controls.emit({
-        type: "INSERT_CELL",
+      this.runtime.controls.insertCell({
+        adjacentCellId: parent.cell.id,
         position: this.insertPosition,
-        id: parent.cell.id,
         data: { cellType },
       });
     }