+++ src/types/core/index.ts
@@ -4,7 +4,8 @@
 
 import { Runtime } from "../runtime";
 import { TemplateResult } from "lit";
-import { CellElement } from "src/components/cell";
+import { CellElement } from "../../components/cell";
+import { StarboardEventInitDict, StarboardEventMap, StarboardEventName } from "../events";
 
 /**
  * Events that can be sent from the cell for central handling in the notebook component.
@@ -102,7 +103,6 @@ export interface NotebookMetadata {
  */
 export interface NotebookContent {
   metadata: NotebookMetadata;
-
   cells: Cell[];
 }
 
@@ -165,11 +165,15 @@ export interface CellElements {
 
 export type IconTemplate = string | ((iconOpts?: { width?: number; height?: number }) => TemplateResult | string);
 
+export type ControlButtonCallbackDispatchFunction<E extends StarboardEventName> = (
+  name: E,
+  detail: StarboardEventInitDict<StarboardEventMap[E]>
+) => boolean;
+
 export interface ControlButton {
   icon: IconTemplate;
   tooltip: string;
-  hide?: boolean;
-  callback: () => any | Promise<any>;
+  callback: (event: Event) => any;
 }
 
 export interface ControlsDefinition {